Renamed app to be more clearer
