+++ backend/game_service/lobby_manager_socket/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class MatchmakingSocketConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'matchmaking_socket'
+    name = 'lobby_manager_socket'
